backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
langgraph
langchain
langchain-openai
pydantic[email]
pydantic-settings
python-dotenv
websockets
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
cryptography
authlib
playwright
beautifulsoup4
gql
aioredis
sentence-transformers
numpy
scikit-learn
stripe
adyen-python-api-library
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
slowapi
python-json-logger
guardrails-ai
bleach
